Move the legacy api calls into a service
